requirements.txt
attrs==17.1.0
opuslib>=1.1.0
PyAudio>=0.2.10
pyaxo>=0.7.3
PyNaCl>=1.0.1
pyperclip>=1.5.27
Twisted[tls]>=16.6.0
txtorcon>=0.19.0
